refactor : 내 관심목록 폴더명 변경 (-> myFavorites)
