+++ src/components/myFavorites/MyFavorites.js
