+++ src/components/myFavorites/table/EnhancedTable.js
